+++ webui/react/src/pages/WebhookList.module.scss
@@ -13,8 +13,4 @@
   h4 {
     margin-bottom: 8;
   }
-  .description {
-    color: var(--theme-colors-monochrome-9);
-    font-size: 14px;
-  }
 }
