+++ webui/react/src/routes/index.tsx
@@ -19,7 +19,7 @@ import TaskList from 'pages/TaskList';
 import { TaskLogsWrapper } from 'pages/TaskLogs';
 import TrialDetails from 'pages/TrialDetails';
 import Wait from 'pages/Wait';
-import Webhooks from 'pages/Webhooks';
+import Webhooks from 'pages/WebhookList';
 import WorkspaceDetails from 'pages/WorkspaceDetails';
 import WorkspaceList from 'pages/WorkspaceList';
 import { RouteConfig } from 'shared/types';
